feat: adiciona modal para confirmacao de exclusao de material
