+++ frontend/src/pages/Material/NewMaterialModal.jsx
@@ -1,52 +1,19 @@
-import { useEffect, useState } from "react";
-import Input from "../Input";
+import { useState } from "react";
+import Input from "../../components/Input";
 import { toast } from "react-toastify";
 import { useAuth } from "../../store/AuthContext";
 
-export default function EditMaterial({ fecharModal, materialId }) {
+export default function NewMaterial({ fecharModal }) {
   const { userData } = useAuth();
   const token = userData?.token;
 
   const [formData, setFormData] = useState({
     nome: "",
     descricao: "",
-    unidade: "",
+    unidade: "un",
   });
 
   const [isSubmitting, setIsSubmitting] = useState(false);
-
-  useEffect(() => {
-    if (materialId) {
-      async function fetchMaterial() {
-        try {
-          const response = await fetch(
-            `http://localhost:8080/api/materiais/${materialId}`,
-            {
-              method: "GET",
-              headers: {
-                Authorization: `Bearer ${token}`,
-              },
-            }
-          );
-
-          if (!response.ok) {
-            throw new Error("Erro ao buscar dados do material.");
-          }
-
-          const data = await response.json();
-          setFormData({
-            nome: data.nome || "",
-            descricao: data.descricao || "",
-            unidade: data.unidade || "",
-          });
-        } catch (error) {
-          toast.error(error.message || "Erro ao buscar dados do material.");
-        }
-      }
-
-      fetchMaterial();
-    }
-  }, [materialId, token]);
 
   function handleFormChange(e) {
     const { name, value } = e.target;
@@ -67,27 +34,25 @@ export default function EditMaterial({ fecharModal, materialId }) {
     setIsSubmitting(true);
 
     try {
-      const response = await fetch(
-        `http://localhost:8080/api/materiais/${materialId}`,
-        {
-          method: "PUT",
-          headers: {
-            "Content-Type": "application/json",
-            Authorization: `Bearer ${token}`,
-          },
-          body: JSON.stringify(formData),
-        }
-      );
+      const response = await fetch("http://localhost:8080/api/materiais", {
+        method: "POST",
+        headers: {
+          "Content-Type": "application/json",
+          Authorization: `Bearer ${token}`,
+        },
+        body: JSON.stringify(formData),
+      });
 
       if (!response.ok) {
         const errorData = await response.json();
-        throw new Error(errorData.message || "Erro ao editar material.");
+        throw new Error(errorData.message || "Erro ao registrar material.");
       }
 
-      toast.success("Material editado com sucesso!");
+      toast.success("Material registrado com sucesso!");
+      setFormData({ nome: "", descricao: "", unidade: "un" });
       fecharModal();
     } catch (error) {
-      toast.error(error.message || "Erro ao editar material.");
+      toast.error(error);
     } finally {
       setIsSubmitting(false);
     }
@@ -96,15 +61,17 @@ export default function EditMaterial({ fecharModal, materialId }) {
   return (
     <div className="max-w-4xl mx-auto px-4 py-8">
       <div className="bg-blue-700 px-6 py-4">
-        <h1 className="text-2xl font-bold text-white">Editar Material</h1>
+        <h1 className="text-2xl font-bold text-white">
+          Registrar Novo Material
+        </h1>
       </div>
 
       <form onSubmit={handleSubmit} className="space-y-4 mt-6">
         <Input
           label="Nome"
+          maxLength="25"
           id="nome"
           type="text"
-          maxLength="25"
           value={formData.nome}
           onChange={handleFormChange}
         />
@@ -117,9 +84,9 @@ export default function EditMaterial({ fecharModal, materialId }) {
         <textarea
           id="descricao"
           name="descricao"
-          value={formData.descricao}
           maxLength={100}
           onChange={handleFormChange}
+          value={formData.descricao}
           className="appearance-none block w-full px-3 py-2 border border-gray-300 rounded-md shadow-sm placeholder-gray-400 focus:outline-none focus:ring-blue-500 focus:border-blue-500"
         />
         <label
@@ -131,13 +98,15 @@ export default function EditMaterial({ fecharModal, materialId }) {
         <select
           id="unidade"
           name="unidade"
-          value={formData.unidade}
           onChange={handleFormChange}
+          value={formData.unidade}
           className="appearance-none block w-full px-3 py-2 border border-gray-300 rounded-md shadow-sm placeholder-gray-400 focus:outline-none focus:ring-blue-500 focus:border-blue-500"
         >
           <option value="kg">kg</option>
           <option value="g">g</option>
-          <option value="un">un</option>
+          <option selected="selected" value="un">
+            un
+          </option>
           <option value="l">l</option>
           <option value="ml">ml</option>
         </select>
@@ -156,7 +125,7 @@ export default function EditMaterial({ fecharModal, materialId }) {
             className="bg-blue-600 hover:bg-blue-700 cursor-pointer rounded px-4 py-2 text-white"
             disabled={isSubmitting}
           >
-            {isSubmitting ? "Salvando..." : "Salvar"}
+            {isSubmitting ? "Adicionando..." : "Adicionar"}
           </button>
         </div>
       </form>
